# GSIE/API/docs/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "LoginRequest": {
        "additionalProperties": false,
        "description": "Requête de login — username + password.",
        "properties": {
          "password": {
            "description": "Mot de passe",
            "maxLength": 500,
            "minLength": 1,
            "title": "Password",
            "type": "string"
          },
          "username": {
            "description": "Nom d'utilisateur",
            "maxLength": 255,
            "minLength": 1,
            "title": "Username",
            "type": "string"
          }
        },
        "required": [
          "username",
          "password"
        ],
        "title": "LoginRequest",
        "type": "object"
      },
      "LogoutRequest": {
        "additionalProperties": false,
        "description": "Requête de logout — refresh token à révoquer.",
        "properties": {
          "refresh_token": {
            "description": "Refresh token JWT à révoquer",
            "minLength": 1,
            "title": "Refresh Token",
            "type": "string"
          }
        },
        "required": [
          "refresh_token"
        ],
        "title": "LogoutRequest",
        "type": "object"
      },
      "LogoutResponse": {
        "additionalProperties": false,
        "description": "Réponse de logout — confirmation de révocation.",
        "properties": {
          "revoked": {
            "title": "Revoked",
            "type": "boolean"
          }
        },
        "required": [
          "revoked"
        ],
        "title": "LogoutResponse",
        "type": "object"
      },
...
      "RefreshRequest": {
        "additionalProperties": false,
        "description": "Requête de refresh — refresh token.",
        "properties": {
          "refresh_token": {
            "description": "Refresh token JWT",
            "minLength": 1,
            "title": "Refresh Token",
            "type": "string"
          }
        },
        "required": [
          "refresh_token"
        ],
        "title": "RefreshRequest",
        "type": "object"
      },
...
      "TokenResponse": {
        "additionalProperties": false,
        "description": "Réponse token — access + refresh.",
        "properties": {
          "access_token": {
            "title": "Access Token",
            "type": "string"
          },
          "expires_in": {
            "description": "Durée de vie du access token en secondes",
            "title": "Expires In",
            "type": "integer"
          },
          "refresh_token": {
            "title": "Refresh Token",
            "type": "string"
          },
          "token_type": {
            "default": "Bearer",
            "title": "Token Type",
            "type": "string"
          }
        },
        "required": [
          "access_token",
          "refresh_token",
          "expires_in"
        ],
        "title": "TokenResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
      "VerifyResponse": {
        "additionalProperties": false,
        "description": "Réponse de vérification — statut du token.",
        "properties": {
          "expires_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires At"
          },
          "subject": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Subject"
          },
          "token_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Token Type"
          },
          "valid": {
            "title": "Valid",
            "type": "boolean"
          }
        },
        "required": [
          "valid"
        ],
        "title": "VerifyResponse",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "GSIE — General System Intelligence Engine API",
    "title": "GSIE API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/auth/login": {
      "post": {
        "description": "Authentifie un utilisateur avec username + password et retourne un access token (15 min) et un refresh token (7 jours). Les tokens sont signés en RS256 (DEC-000019).",
        "operationId": "login_api_v1_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Authentifier un utilisateur",
        "tags": [
          "auth"
        ]
      }
    },
...
    "/api/v1/auth/logout": {
      "post": {
        "description": "Révoque le refresh token fourni dans le registre. Le access token actif reste valide jusqu'à son expiration (15 min) — c'est la limite du JWT sans liste noire centralisée. Le refresh token ne peut plus être utilisé pour obtenir un nouveau access token.",
        "operationId": "logout_api_v1_auth_logout_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LogoutRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LogoutResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Déconnexion — révocation du refresh token",
        "tags": [
          "auth"
        ]
      }
    },
...
    "/api/v1/auth/refresh": {
      "post": {
        "description": "Échange un refresh token valide contre un nouveau access token et un nouveau refresh token (rotation).",
        "operationId": "refresh_token_api_v1_auth_refresh_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Rafraîchir le token d'accès",
        "tags": [
          "auth"
        ]
      }
    },
...
    "/api/v1/auth/verify": {
      "get": {
        "description": "Vérifie la validité du token d'accès fourni dans le header Authorization.",
        "operationId": "verify_access_token_api_v1_auth_verify_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VerifyResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Vérifier un token",
        "tags": [
          "auth"
        ]
      }
    },
...
  },
...
}